Prefer operator<<() to pretty()
It is more idiomatic, we didn't used it
in the past because Position::pretty(Move)
had a calling argument, but now we can.

As an added benefit, we avoid a lot of string
copies in the process because now we avoid
std::ostringstream ss.

No functional change.

(cherry picked from commit 5644e14d0e3d69b3f845e475771564ffb3e25445)

# Conflicts:
#	src/position.cpp
#	src/position.h