Added "mindepth" "go" option.
This allows for a command like this:

    go mindepth 5 movetime 1000 depth 10

This will at least search 5 plys, search no more than 10 plys, and search no longer than 1 second. (It will stop when either movetime or depth parameters are met provided that it is greater than 5 plys deep.)

Note that this will produce a move with at least depth 6 (it finishes 5 plys).

Note that "mindepth" only takes effect when using "movetime".

Also, note that Stockfish considers "depth" as max depth. So if "mindepth" is greater than "depth", "depth" will be used.